refactor: Move files to correct feature package
This commit moves files related to the list and setting features to their respective feature packages.

- ListViewModel.
kt is moved from feature/main to feature/list.
- SettingNavigation.kt is moved from feature/main/navigation to feature/setting/navigation.
- SettingScreen.kt is moved from feature/main to feature/setting.
- ListNavigation.kt is moved from feature/main/navigation
 to feature/list/navigation.
- SettingViewModel.kt is moved from feature/main to feature/setting.
- ListScreen.kt is moved from feature/main to feature/list.

This change improves the organization of the codebase and makes it easier to maintain.
