+++ feature/list/src/main/kotlin/io/github/shinhyo/brba/feature/list/navigation/ListNavigation.kt
@@ -13,19 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.shinhyo.brba.feature.main.navigation
+package io.github.shinhyo.brba.feature.list.navigation
 
+import androidx.compose.animation.AnimatedVisibilityScope
+import androidx.compose.animation.SharedTransitionScope
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navOptions
-import io.github.shinhyo.brba.feature.main.SettingRoute
+import io.github.shinhyo.brba.core.model.BrbaCharacter
+import io.github.shinhyo.brba.feature.list.ListRoute
 
-const val SETTING_ROUTE = "setting_route"
+const val LIST_ROUTE = "list_route"
 
-fun NavController.navigateSetting() {
+fun NavController.navigateList() {
     navigate(
-        route = SETTING_ROUTE,
+        route = LIST_ROUTE,
         navOptions = navOptions {
             popUpTo(graph.startDestinationId) {
                 saveState = true
@@ -36,10 +39,17 @@ fun NavController.navigateSetting() {
     )
 }
 
-fun NavGraphBuilder.settingComposable() {
+context(SharedTransitionScope)
+fun NavGraphBuilder.listComposable(
+    navigateToDetail: (BrbaCharacter) -> Unit,
+    animatedVisibilityScope: AnimatedVisibilityScope,
+) {
     composable(
-        route = SETTING_ROUTE,
+        route = LIST_ROUTE,
     ) {
-        SettingRoute()
+        ListRoute(
+            navigateToDetail = navigateToDetail,
+            animatedVisibilityScope = animatedVisibilityScope,
+        )
     }
 }
