+++ feature/setting/src/main/kotlin/io/github/shinhyo/brba/feature/setting/SettingViewModel.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.shinhyo.brba.feature.main
+package io.github.shinhyo.brba.feature.setting
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
